+++ tests/testthat/test-oauth-flow-auth-code.R
@@ -75,12 +75,12 @@ test_that("old args are deprecated", {
   expect_snapshot(
     redirect <- normalize_redirect_uri("http://localhost", port = 1234)
   )
-  expect_equal(redirect$uri, "http://localhost:1234")
+  expect_equal(redirect$uri, "http://localhost:1234/")
 
   expect_snapshot(
     redirect <- normalize_redirect_uri("http://x.com", host_name = "y.com")
   )
-  expect_equal(redirect$uri, "http://y.com")
+  expect_equal(redirect$uri, "http://y.com/")
 
   expect_snapshot(
     redirect <- normalize_redirect_uri("http://x.com", host_ip = "y.com")